Combining three components into one.
