--- src/components/TableTrees.tsx
@@ -1,39 +1,0 @@
-import React, { useState } from 'react';
-
-
-export function TableTrees() {
-    const renderTable = () => {
-        const rows = [];
-        const cellContents = ['Yellow Trees', 'Green Trees', '1', '4', '2', '3', '3', '2', '4', '?'];
-    
-        for (let i = 0; i < 5; i++) {
-        const cells = [];
-        for (let j = 0; j < 2; j++) {
-            const index = i * 2 + j; 
-            cells.push(
-            <td key={j} className="border border-black">
-                <div className="flex flex-wrap justify-center space-x-4">
-                <div className="text-white p-2 rounded-md shadow-sm">
-                    {index < cellContents.length ? cellContents[index] : '?'}
-                </div>
-                </div>
-            </td>
-            );
-        }
-        rows.push(<tr key={i}>{cells}</tr>);
-        }
-  
-    return (
-      <table className="border-collapse border bg-primary-green border-black">
-        <tbody>{rows}</tbody>
-      </table>
-    );
-  };
-
-return (
-  <div className="flex flex-col items-center justify-center space-y-4">
-      {renderTable()}
-      </div>
-  );
-}
-
